refact: useOutsideClick 커스텀 훅 구현
#76
