+++ src/hooks/useOutsideClick.tsx
@@ -1,0 +1,19 @@
+import { useEffect } from "react";
+
+type OutSideClickProps = {
+  modalRef: React.RefObject<HTMLDivElement>;
+  onClose: () => void;
+};
+
+export function useOutsideClick({ modalRef, onClose }: OutSideClickProps) {
+  function handleClose(e: MouseEvent) {
+    if (modalRef.current && !modalRef.current.contains(e.target as Element)) onClose();
+  }
+
+  useEffect(() => {
+    window.addEventListener("mousedown", handleClose);
+    return () => {
+      window.removeEventListener("mousedown", handleClose);
+    };
+  }, []);
+}
